+++ regression/scripts/AS2/mission_AS2.py
@@ -24,7 +24,6 @@ Data, Container,
 )
 
 # the analysis functions
-from the_aircraft_function import the_aircraft_function
 from SUAVE.Methods.Propulsion.turbojet_sizing import turbojet_sizing
 
 
@@ -169,441 +168,7 @@ def base_analysis(vehicle):
     # done!
     return analyses    
 
-# ----------------------------------------------------------------------
-#   Define the Vehicle
-# ----------------------------------------------------------------------
 
-def vehicle_setup():
-    
-    # ------------------------------------------------------------------
-    #   Initialize the Vehicle
-    # ------------------------------------------------------------------    
-    
-    vehicle = SUAVE.Vehicle()
-    vehicle.tag = 'AS2'    
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle-level Properties
-    # ------------------------------------------------------------------    
-
-    # mass properties
-    vehicle.mass_properties.max_takeoff               = 52163.   # kg
-    vehicle.mass_properties.operating_empty           = 22500.   # kg
-    vehicle.mass_properties.takeoff                   = 52163.   # kg
-    vehicle.mass_properties.cargo                     = 1000.  * Units.kilogram   
-    
-    vehicle.mass_properties.center_of_gravity         = [26.3 * Units.feet, 0, 0] 
-    vehicle.mass_properties.moments_of_inertia.tensor = [[10 ** 5, 0, 0],[0, 10 ** 6, 0,],[0,0, 10 ** 7]] # Not Correct
-    
-    # envelope properties
-    vehicle.envelope.ultimate_load = 3.5
-    vehicle.envelope.limit_load    = 1.5
-
-    # basic parameters
-    vehicle.reference_area         = 125.4      
-    vehicle.passengers             = 8
-    vehicle.systems.control        = "fully powered" 
-    vehicle.systems.accessories    = "long range"
-    
-    
-    # ------------------------------------------------------------------        
-    #   Main Wing
-    # ------------------------------------------------------------------        
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'main_wing'
-    
-    wing.aspect_ratio            = 3.63
-    wing.sweeps.quarter_chord    = 0 * Units.deg
-    wing.thickness_to_chord      = 0.03
-    wing.taper                   = 0.7
-    wing.span_efficiency         = 0.74
-    
-    wing.spans.projected         = 21.0    
-    
-    wing.chords.root             = 12.9
-    wing.chords.tip              = 1.0
-    wing.chords.mean_aerodynamic = 7.0
-    
-    wing.areas.reference         = 125.4 
-    
-    wing.twists.root             = 0.0 * Units.degrees
-    wing.twists.tip              = 2.0 * Units.degrees
-    
-    wing.origin                  = [20,0,0]
-    wing.aerodynamic_center      = [5,0,0] 
-    
-    wing.vertical                = False
-    wing.symmetric               = True
-    wing.high_lift               = True
-    wing.high_mach               = True
-    wing.transition_x_upper      = 0.9
-    wing.transition_x_lower      = 0.9
-    
-    wing.dynamic_pressure_ratio  = 1.0
-    
-    # add to vehicle
-    vehicle.append_component(wing)
-
-
-    # ------------------------------------------------------------------        
-    #  Horizontal Stabilizer
-    # ------------------------------------------------------------------        
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'horizontal_stabilizer'
-    
-    wing.aspect_ratio            = 2.0      #
-    wing.sweeps.quarter_chord    = 0 * Units.deg
-    wing.thickness_to_chord      = 0.03
-    wing.taper                   = 0.5
-    wing.span_efficiency         = 0.74
-    
-    wing.spans.projected         = 7.0      #
-
-    wing.chords.root             = 4.0 
-    wing.chords.tip              = 2.0    
-    wing.chords.mean_aerodynamic = 3.3
-
-    wing.areas.reference         = 24.5    #
-    
-    wing.twists.root             = 3.0 * Units.degrees
-    wing.twists.tip              = 3.0 * Units.degrees  
-    
-    wing.origin                  = [46,0,0]
-    wing.aerodynamic_center      = [2,0,0]
-    
-    wing.vertical                = False 
-    wing.symmetric               = True
-    wing.high_mach               = True
-    wing.transition_x_upper      = 0.9
-    wing.transition_x_lower      = 0.9    
-    
-    wing.dynamic_pressure_ratio  = 0.9  
-    
-    # add to vehicle
-    vehicle.append_component(wing)
-    
-    
-    # ------------------------------------------------------------------
-    #   Vertical Stabilizer
-    # ------------------------------------------------------------------
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'vertical_stabilizer'    
-    
-    wing.aspect_ratio            = 1.3      #
-    wing.sweeps.quarter_chord    = 45 * Units.deg
-    wing.thickness_to_chord      = 0.08
-    wing.taper                   = 0.5
-    wing.span_efficiency         = 0.9
-    
-    wing.spans.projected         = 3.5      #    
-
-    wing.chords.root             = 6.0
-    wing.chords.tip              = 3.0
-    wing.chords.mean_aerodynamic = 4.0
-    
-    wing.areas.reference         = 33.9    #
-    
-    wing.twists.root             = 0.0 * Units.degrees
-    wing.twists.tip              = 0.0 * Units.degrees  
-    
-    wing.origin                  = [40,0,0]
-    wing.aerodynamic_center      = [2,0,0]    
-    
-    wing.vertical                = True 
-    wing.symmetric               = False
-    wing.t_tail                  = False
-    wing.high_mach               = True
-    wing.transition_x_upper      = 0.9
-    wing.transition_x_lower      = 0.9      
-    
-    wing.dynamic_pressure_ratio  = 1.0
-        
-    # add to vehicle
-    vehicle.append_component(wing)
-
-
-    # ------------------------------------------------------------------
-    #  Fuselage
-    # ------------------------------------------------------------------
-    
-    fuselage = SUAVE.Components.Fuselages.Fuselage()
-    fuselage.tag = 'fuselage'
-    
-    #fuselage.number_coach_seats    = vehicle.passengers
-    fuselage.seats_abreast         = 2
-    fuselage.seat_pitch            = 1
-    
-    fuselage.fineness.nose         = 4.0 # These finenesses are smaller than the real value due to limitations of existing functions
-    fuselage.fineness.tail         = 4.0
-    
-    fuselage.lengths.nose          = 12.
-    fuselage.lengths.tail          = 25.5
-    fuselage.lengths.cabin         = 11.5
-    fuselage.lengths.total         = 49.    
-    fuselage.lengths.fore_space    = 16.3
-    fuselage.lengths.aft_space     = 16.3  
-    
-    fuselage.width                 = 2.35
-    
-    fuselage.heights.maximum       = 2.55    #
-    fuselage.heights.at_quarter_length          = 4. # Not correct
-    fuselage.heights.at_three_quarters_length   = 4. # Not correct
-    fuselage.heights.at_wing_root_quarter_chord = 4. # Not correct
-
-    fuselage.areas.side_projected  = 4.* 59.8 #  Not correct
-    fuselage.areas.wetted          = 615.
-    fuselage.areas.front_projected = 5.1
-    
-    fuselage.effective_diameter    = 2.4
-    
-    fuselage.differential_pressure = 50**5 * Units.pascal    # Maximum differential pressure
-    
-    # add to vehicle
-    vehicle.append_component(fuselage)
-    
-        
-    # ------------------------------------------------------------------
-    #   Turbojet Network
-    # ------------------------------------------------------------------    
-    
-    #instantiate the gas turbine network
-    turbojet = SUAVE.Components.Energy.Networks.Turbojet_Super()
-    turbojet.tag = 'turbojet'
-    
-    # setup
-    turbojet.number_of_engines = 3.0
-    turbojet.engine_length     = 8.0
-    turbojet.nacelle_diameter  = 1.580
-    
-    # working fluid
-    turbojet.working_fluid = SUAVE.Attributes.Gases.Air()
-    
-    
-    # ------------------------------------------------------------------
-    #   Component 1 - Ram
-    
-    # to convert freestream static to stagnation quantities
-    
-    # instantiate
-    ram = SUAVE.Components.Energy.Converters.Ram()
-    ram.tag = 'ram'
-    
-    # add to the network
-    turbojet.append(ram)
-
-
-    # ------------------------------------------------------------------
-    #  Component 2 - Inlet Nozzle
-    
-    # instantiate
-    inlet_nozzle = SUAVE.Components.Energy.Converters.Compression_Nozzle()
-    inlet_nozzle.tag = 'inlet_nozzle'
-    
-    # setup
-    inlet_nozzle.polytropic_efficiency = 1.0
-    inlet_nozzle.pressure_ratio        = 1.0
-    
-    # add to network
-    turbojet.append(inlet_nozzle)
-    
-    
-    # ------------------------------------------------------------------
-    #  Component 3 - Low Pressure Compressor
-    
-    # instantiate 
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'low_pressure_compressor'
-
-    # setup
-    compressor.polytropic_efficiency = 1.0
-    compressor.pressure_ratio        = 5.0    
-    
-    # add to network
-    turbojet.append(compressor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 4 - High Pressure Compressor
-    
-    # instantiate
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'high_pressure_compressor'
-    
-    # setup
-    compressor.polytropic_efficiency = 1.0
-    compressor.pressure_ratio        = 10.0  
-    
-    # add to network
-    turbojet.append(compressor)
-
-
-    # ------------------------------------------------------------------
-    #  Component 5 - Low Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='low_pressure_turbine'
-    
-    # setup
-    turbine.mechanical_efficiency = 1.0
-    turbine.polytropic_efficiency = 1.0     
-    
-    # add to network
-    turbojet.append(turbine)
-    
-      
-    # ------------------------------------------------------------------
-    #  Component 6 - High Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='high_pressure_turbine'
-
-    # setup
-    turbine.mechanical_efficiency = 1.0
-    turbine.polytropic_efficiency = 1.0     
-    
-    # add to network
-    turbojet.append(turbine)
-      
-    
-    # ------------------------------------------------------------------
-    #  Component 7 - Combustor
-    
-    # instantiate    
-    combustor = SUAVE.Components.Energy.Converters.Combustor()   
-    combustor.tag = 'combustor'
-    
-    # setup
-    combustor.efficiency                = 1.0
-    combustor.alphac                    = 1.0     
-    combustor.turbine_inlet_temperature = 1500.
-    combustor.pressure_ratio            = 1.0
-    combustor.fuel_data                 = SUAVE.Attributes.Propellants.Jet_A()    
-    
-    # add to network
-    turbojet.append(combustor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 8 - Core Nozzle
-    
-    # instantiate
-    nozzle = SUAVE.Components.Energy.Converters.Supersonic_Nozzle()  
-    #nozzle = SUAVE.Components.Energy.Converters.Expansion_Nozzle()
-    nozzle.tag = 'core_nozzle'
-    
-    # setup
-    nozzle.polytropic_efficiency = 1.0
-    nozzle.pressure_ratio        = 1.0    
-    
-    # add to network
-    turbojet.append(nozzle)
-    
-    # ------------------------------------------------------------------
-    #  Component 9 - Divergening Nozzle
-    
-    
-    
-    
-    # ------------------------------------------------------------------
-    #Component 10 : thrust (to compute the thrust)
-    thrust = SUAVE.Components.Energy.Processes.Thrust()       
-    thrust.tag ='compute_thrust'
- 
-    #total design thrust (includes all the engines)
-    thrust.total_design             = 2*15000. * Units.N #Newtons
- 
-    #design sizing conditions
-    altitude      = 35000.0*Units.ft
-    mach_number   = 2.02
-    isa_deviation = 0.
-    
-    # add to network
-    turbojet.thrust = thrust
-
-    #size the turbojet
-    turbojet_sizing(turbojet,mach_number,altitude)   
-    
-    # add  gas turbine network gt_engine to the vehicle
-    vehicle.append_component(turbojet)      
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle Definition Complete
-    # ------------------------------------------------------------------
-
-    return vehicle
-
-
-# ----------------------------------------------------------------------
-#   Define the Configurations
-# ---------------------------------------------------------------------
-
-def configs_setup(vehicle):
-    
-    # ------------------------------------------------------------------
-    #   Initialize Configurations
-    # ------------------------------------------------------------------
-    
-    configs = SUAVE.Components.Configs.Config.Container()
-    
-    base_config = SUAVE.Components.Configs.Config(vehicle)
-    base_config.tag = 'base'
-    configs.append(base_config)
-    
-    # ------------------------------------------------------------------
-    #   Cruise Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'cruise'
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Takeoff Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'takeoff'
-    
-    config.wings['main_wing'].flaps.angle = 20. * Units.deg
-    config.wings['main_wing'].slats.angle = 25. * Units.deg
-    
-    config.V2_VS_ratio = 1.21
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Landing Configuration
-    # ------------------------------------------------------------------
-
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'landing'
-    
-    config.wings['main_wing'].flaps_angle = 30. * Units.deg
-    config.wings['main_wing'].slats_angle = 25. * Units.deg
-
-    config.Vref_VS_ratio = 1.23
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    
-    # done!
-    return configs
-
-# ----------------------------------------------------------------------
-#   Sizing for the Vehicle Configs
-# ----------------------------------------------------------------------
 
 # ----------------------------------------------------------------------
 #   Plot Mission
